ssh: Skip test if passwordless authentication to localhost not possible.
Fixes commit dcc96d324ff7c8add9da1bbb1412586fe6f4b505.

+++ tests/test-ssh.sh
@@ -38,6 +38,14 @@ set -x
 requires sshd -t -f ssh/sshd_config
 requires qemu-img --version
 
+# Check that ssh to localhost will work without any passwords or phrases.
+requires ssh -V
+if ! ssh -o PreferredAuthentications=none,publickey localhost echo </dev/null
+then
+    echo "$0: passwordless/phraseless authentication to localhost not possible"
+    exit 77
+fi
+
 files="ssh.img"
 rm -f $files
 cleanup_fn rm -f $files
